+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # druxt-test-utils
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [45bc0b9]
+  - druxt@0.14.0
+
 ## 0.2.3
 
 ### Patch Changes